move annotator node to its own module
