+++ tests/modelgauge_tests/annotator_tests/composer_tests/test_annotator_node.py
@@ -4,8 +4,8 @@ import pytest
 
 from modelgauge.annotation import SafetyAnnotation
 from modelgauge.annotator import Annotator
+from modelgauge.annotators.composer.annotator_node import AnnotatorNode
 from modelgauge.annotators.composer.context import EvalContext
-from modelgauge.annotators.composer.nodes import AnnotatorNode
 
 
 @pytest.fixture
